requirements.txt
Flask==3.0.0
Flask-MySQLdb==1.0.1
Werkzeug==3.0.1
mysqlclient==2.2.0
